backend/requirements.txt
fastapi
uvicorn
openai
google-generativeai
python-dotenv
python-multipart
pydub
librosa
numpy
SpeechRecognition